% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{error_ests}
\alias{error_ests}
\alias{error_ests<-}
\title{generic error_ests getter}
\usage{
error_ests(x)

error_ests(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{tibble of error parameter estimates}
}
\description{
generic error_ests getter

generic error_ests setter
}
