% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{fit_summary}
\alias{fit_summary}
\alias{fit_summary<-}
\title{generic fit_summary getter}
\usage{
fit_summary(x)

fit_summary(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{character string description of fit}
}
\description{
generic fit_summary getter

generic fit_summary setter
}
