% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde-package.R
\docType{package}
\name{hmde-package}
\alias{hmde-package}
\alias{hmde}
\title{The 'hmde' package.}
\description{
A package to implement a selection of hierarchical Bayesian longitudinal models for inverse Bayesian problems.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.23. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://traitecoevo.github.io/hmde/}
  \item Report bugs at \url{https://github.com/traitecoevo/hmde/issues}
}

}
\author{
\strong{Maintainer}: Tess O'Brien \email{tesslacoil@protonmail.com} (\href{https://orcid.org/0009-0000-5821-8801}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID}) [contributor]
  \item David Warton \email{david.warton@unsw.edu.au} (\href{https://orcid.org/0000-0002-1642-628X}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Fonti Kar \email{f.kar@unsw.edu.au} (\href{https://orcid.org/0000-0002-2760-3974}{ORCID}) [contributor]
}

}
