% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_affine_de}
\alias{hmde_affine_de}
\title{Differential equation for affine growth single individual model}
\usage{
hmde_affine_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real}

\item{pars}{list of parameters beta_0, beta_1}
}
\value{
value of differential equation at y
}
\description{
Differential equation for affine growth single individual model
}
