% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_assign_data.R
\name{hmde_assign_data}
\alias{hmde_assign_data}
\title{Assign data to template}
\usage{
hmde_assign_data(model_template, ...)
}
\arguments{
\item{model_template}{output from hmde_model}

\item{...}{data-masking name-value pairs}
}
\value{
updated named list with your data assigned to Stan model parameters
}
\description{
Assign data to template
}
