% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_canham_de}
\alias{hmde_canham_de}
\title{Differential equation for Canham growth single and multi- individual models}
\usage{
hmde_canham_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real}

\item{pars}{list of parametera g_max, S_max, k}
}
\value{
value of differential equation at y
}
\description{
Differential equation for Canham growth single and multi- individual models
}
