% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_const_de}
\alias{hmde_const_de}
\title{Differential equation for constant growth single and multi- individual models}
\usage{
hmde_const_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real}

\item{pars}{list of parameter beta}
}
\value{
value of differential equation at y
}
\description{
Differential equation for constant growth single and multi- individual models
}
