% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\docType{class}
\name{hmde_data_template-class}
\alias{hmde_data_template-class}
\alias{hmde_data_template}
\alias{validator for hmde_data_template
Validation function for hmde_data_template class.}
\alias{model_name,hmde_data_template-method}
\alias{model_name<-,hmde_data_template-method}
\alias{model_level,hmde_data_template-method}
\alias{model_level<-,hmde_data_template-method}
\alias{obs_data,hmde_data_template-method}
\alias{obs_data<-,hmde_data_template-method}
\alias{prior_pars,hmde_data_template-method}
\alias{prior_pars<-,hmde_data_template-method}
\alias{par_names,hmde_data_template-method}
\alias{par_names<-,hmde_data_template-method}
\alias{show,hmde_data_template-method}
\alias{print,hmde_data_template-method}
\alias{summary,hmde_data_template-method}
\alias{plot,hmde_data_template,ANY-method}
\title{An S4 class to prepare input data for a hmde model.}
\usage{
\S4method{model_name}{hmde_data_template}(x)

\S4method{model_name}{hmde_data_template}(x) <- value

\S4method{model_level}{hmde_data_template}(x)

\S4method{model_level}{hmde_data_template}(x) <- value

\S4method{obs_data}{hmde_data_template}(x)

\S4method{obs_data}{hmde_data_template}(x) <- value

\S4method{prior_pars}{hmde_data_template}(x)

\S4method{prior_pars}{hmde_data_template}(x) <- value

\S4method{par_names}{hmde_data_template}(x)

\S4method{par_names}{hmde_data_template}(x) <- value

hmde_data_template(
  model_name,
  obs_data = NULL,
  prior_pars = NULL,
  model_level = NA_character_,
  ...
)

\S4method{show}{hmde_data_template}(object)

\S4method{print}{hmde_data_template}(x)

\S4method{summary}{hmde_data_template}(object)

\S4method{plot}{hmde_data_template,ANY}(x)
}
\arguments{
\item{x}{hmde_data_template class object}

\item{value}{list}

\item{model_name}{character string name of a hmde model}

\item{obs_data}{list or tibble containing observational data vectors}

\item{prior_pars}{list containing prior parameters}

\item{model_level}{character string specifying whether single or multiple inds}

\item{...}{data-masking name-value pairs allowing specific input of elements}

\item{object}{hmde_data_template class object}
}
\value{
hmde_data_template class object
}
\description{
An S4 class to prepare input data for a hmde model.

model_name getter
Getter for model_name in hmde_data_template object

model_name setter
Setter for model_name in hmde_data_template object

model_level getter
Getter for model_level in hmde_data_template object

model_level setter
Setter for model_level in hmde_data_template object

obs_data getter
Getter for obs_data in hmde_data_template object

obs_data setter
Setter for obs_data in hmde_data_template object

prior_pars getter
Getter for prior_pars in hmde_data_template object

prior_pars setter
Setter for prior_pars in hmde_data_template object

par_names getter
Getter for par_names in hmde_data_template object

par_names setter
Setter for par_names in hmde_data_template object.

Constructor
Constructor function for hmde_data_template class, with data assignment ready for model fitting.

Show function for hmde_data_template object

Print function for hmde_data_template object

Summary function for hmde_data_template object

Plot function for hmde_data_template object
}
\section{Slots}{

\describe{
\item{\code{model_name}}{character string}

\item{\code{model_level}}{character string}

\item{\code{obs_data}}{list of input data}

\item{\code{prior_pars}}{list of prior parameters}

\item{\code{par_names}}{list of model paramter names at each level}
}}

\examples{
# basic usage of hmde_data_template
hmde_data_template("constant_single_ind")

# basic usage of hmde_data_template with data input
hmde_data_template("constant_single_ind",
  obs_data = list(
    n_obs = 3,
    y_obs = c(1,1,1),
    obs_index = 1:3,
    time = 0:2
  )
)

#basic usage of hmde_data_template with prior input
hmde_data_template("constant_single_ind",
  prior_pars = list(
    prior_pars_ind_beta = c(1,2),
    prior_pars_global_error_sigma = c(1,3)
  )
)

}
