% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\docType{class}
\name{hmde_estimates-class}
\alias{hmde_estimates-class}
\alias{model_name,hmde_estimates-method}
\alias{model_name<-,hmde_estimates-method}
\alias{model_level,hmde_estimates-method}
\alias{model_level<-,hmde_estimates-method}
\alias{method,hmde_estimates-method}
\alias{method<-,hmde_estimates-method}
\alias{runtime,hmde_estimates-method}
\alias{runtime<-,hmde_estimates-method}
\alias{fit_summary,hmde_estimates-method}
\alias{fit_summary<-,hmde_estimates-method}
\alias{measurement_ests,hmde_estimates-method}
\alias{measurement_ests<-,hmde_estimates-method}
\alias{individual_ests,hmde_estimates-method}
\alias{individual_ests<-,hmde_estimates-method}
\alias{population_ests,hmde_estimates-method}
\alias{population_ests<-,hmde_estimates-method}
\alias{error_ests,hmde_estimates-method}
\alias{error_ests<-,hmde_estimates-method}
\alias{prior_pars,hmde_estimates-method}
\alias{prior_pars<-,hmde_estimates-method}
\alias{par_names,hmde_estimates-method}
\alias{par_names<-,hmde_estimates-method}
\alias{hmde_estimates}
\alias{show,hmde_estimates-method}
\alias{print,hmde_estimates-method}
\alias{summary,hmde_estimates-method}
\alias{plot,hmde_estimates,ANY-method}
\title{hmde_estimates class object
An S4 class to contain model information and parameter estimates for a hmde model.}
\usage{
\S4method{model_name}{hmde_estimates}(x)

\S4method{model_name}{hmde_estimates}(x) <- value

\S4method{model_level}{hmde_estimates}(x)

\S4method{model_level}{hmde_estimates}(x) <- value

\S4method{method}{hmde_estimates}(x)

\S4method{method}{hmde_estimates}(x) <- value

\S4method{runtime}{hmde_estimates}(x)

\S4method{runtime}{hmde_estimates}(x) <- value

\S4method{fit_summary}{hmde_estimates}(x)

\S4method{fit_summary}{hmde_estimates}(x) <- value

\S4method{measurement_ests}{hmde_estimates}(x)

\S4method{measurement_ests}{hmde_estimates}(x) <- value

\S4method{individual_ests}{hmde_estimates}(x)

\S4method{individual_ests}{hmde_estimates}(x) <- value

\S4method{population_ests}{hmde_estimates}(x)

\S4method{population_ests}{hmde_estimates}(x) <- value

\S4method{error_ests}{hmde_estimates}(x)

\S4method{error_ests}{hmde_estimates}(x) <- value

\S4method{prior_pars}{hmde_estimates}(x)

\S4method{prior_pars}{hmde_estimates}(x) <- value

\S4method{par_names}{hmde_estimates}(x)

\S4method{par_names}{hmde_estimates}(x) <- value

hmde_estimates(fit, obs_data)

\S4method{show}{hmde_estimates}(object)

\S4method{print}{hmde_estimates}(x)

\S4method{summary}{hmde_estimates}(object)

\S4method{plot}{hmde_estimates,ANY}(x)
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{vector}

\item{fit}{stanfit class object}

\item{obs_data}{tbl_df class object with variables for ind_id, time, y_obs}
}
\value{
hmde_estimates class object
}
\description{
Getter for par_names in hmde_estimates object
}
\section{Slots}{

\describe{
\item{\code{model_name}}{name of the hmde model}

\item{\code{model_level}}{whether the model functions at the single or multi-ind level}

\item{\code{method}}{sampling method eg. MCMC}

\item{\code{runtime}}{matrix of chain runtime}

\item{\code{fit_summary}}{description of the fit}

\item{\code{measurement_ests}}{tibble of measurement-level estimates}

\item{\code{individual_ests}}{tibble of individual-level estimates}

\item{\code{population_ests}}{list of population_level estimates}

\item{\code{error_ests}}{tibble of error parameter estimates}

\item{\code{prior_pars}}{list of prior parameters}

\item{\code{par_names}}{list of model paramter names at each level}
}}

\examples{
# basic usage of hmde_estimates
hmde_data_template("constant_single_ind",
                   obs_data = Trout_Size_Data[1:4,]) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_estimates(obs_data = Trout_Size_Data[1:4,])

}
