% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_extract_Rhat.R
\name{hmde_extract_Rhat}
\alias{hmde_extract_Rhat}
\title{Calculate Rhat statistics for a hmde_fit object}
\usage{
hmde_extract_Rhat(fit)
}
\arguments{
\item{fit}{hmde_fit fitted model object, output of hmde_run}
}
\value{
named vector of Rhat values
}
\description{
Calculate Rhat statistics for a hmde_fit object
}
\examples{
# basic usage of hmde_extract_Rhat
hmde_data_template("constant_single_ind",
                   obs_data = Trout_Size_Data[1:4,])|>
  hmde_run(chains = 2, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_extract_Rhat()

}
