% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\docType{class}
\name{hmde_fit-class}
\alias{hmde_fit-class}
\alias{hmde_fit}
\alias{data_template-validator}
\alias{model_level,hmde_fit-method}
\alias{model_level<-,hmde_fit-method}
\alias{obs_data,hmde_fit-method}
\alias{obs_data<-,hmde_fit-method}
\alias{prior_pars,hmde_fit-method}
\alias{prior_pars<-,hmde_fit-method}
\alias{par_names,hmde_fit-method}
\alias{par_names<-,hmde_fit-method}
\alias{model_name,hmde_fit-method}
\alias{model_name<-,hmde_fit-method}
\alias{model_pars,hmde_fit-method}
\alias{model_pars<-,hmde_fit-method}
\alias{par_dims,hmde_fit-method}
\alias{par_dims<-,hmde_fit-method}
\alias{mode,hmde_fit-method}
\alias{mode<-,hmde_fit-method}
\alias{sim,hmde_fit-method}
\alias{sim<-,hmde_fit-method}
\alias{inits,hmde_fit-method}
\alias{inits<-,hmde_fit-method}
\alias{stan_args,hmde_fit-method}
\alias{stan_args<-,hmde_fit-method}
\alias{stanmodel,hmde_fit-method}
\alias{stanmodel<-,hmde_fit-method}
\alias{.MISC,hmde_fit-method}
\alias{.MISC<-,hmde_fit-method}
\title{hmde_fit class
An S4 class that inherits the class structure of a stanfit object, with the addition of slots to contain input data and prior information.}
\usage{
\S4method{model_level}{hmde_fit}(x)

\S4method{model_level}{hmde_fit}(x) <- value

\S4method{obs_data}{hmde_fit}(x)

\S4method{obs_data}{hmde_fit}(x) <- value

\S4method{prior_pars}{hmde_fit}(x)

\S4method{prior_pars}{hmde_fit}(x) <- value

\S4method{par_names}{hmde_fit}(x)

\S4method{par_names}{hmde_fit}(x) <- value

\S4method{model_name}{hmde_fit}(x)

\S4method{model_name}{hmde_fit}(x) <- value

\S4method{model_pars}{hmde_fit}(x)

\S4method{model_pars}{hmde_fit}(x) <- value

\S4method{par_dims}{hmde_fit}(x)

\S4method{par_dims}{hmde_fit}(x) <- value

\S4method{mode}{hmde_fit}(x)

\S4method{mode}{hmde_fit}(x) <- value

\S4method{sim}{hmde_fit}(x)

\S4method{sim}{hmde_fit}(x) <- value

\S4method{inits}{hmde_fit}(x)

\S4method{inits}{hmde_fit}(x) <- value

\S4method{stan_args}{hmde_fit}(x)

\S4method{stan_args}{hmde_fit}(x) <- value

\S4method{stanmodel}{hmde_fit}(x)

\S4method{stanmodel}{hmde_fit}(x) <- value

\S4method{.MISC}{hmde_fit}(x)

\S4method{.MISC}{hmde_fit}(x) <- value
}
\arguments{
\item{x}{hmde_fit class object}

\item{value}{environment}

\item{object}{hmde_fit class object}
}
\description{
Validation function for hmde_fit class.
}
\section{Slots}{

\describe{
\item{\code{model_level}}{character string}

\item{\code{obs_data}}{list of input data}

\item{\code{prior_pars}}{list of prior parameters}

\item{\code{par_names}}{list of model paramter names at each level}

\item{\code{model_name}}{character string inherited from stanfit}

\item{\code{model_pars}}{character vector inherited from stanfit}

\item{\code{par_dims}}{list inherited from stanfit}

\item{\code{mode}}{integer inherited from stanfit}

\item{\code{sim}}{list inherited from stanfit}

\item{\code{inits}}{list inherited from stanfit}

\item{\code{stan_args}}{list inherited from stanfit}

\item{\code{stanmodel}}{stanmodel object inherited from stanfit}

\item{\code{date}}{character inherited from stanfit}

\item{\code{.MISC}}{environment inherited from stanfit}
}}

\keyword{internal}
