% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model.R
\name{hmde_model}
\alias{hmde_model}
\title{helper function for hmde_data_template that provides data structure for model,
Also used to guide the user on the required structure for the model they want.}
\usage{
hmde_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
hmde_data_template class that suits Stan model input
}
\description{
helper function for hmde_data_template that provides data structure for model,
Also used to guide the user on the required structure for the model they want.
}
\examples{
# basic usage of hmde_model
hmde_model("constant_single_ind")

}
