% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_model_des}
\alias{hmde_model_des}
\title{Function to select DE given model name}
\usage{
hmde_model_des(model = NULL)
}
\arguments{
\item{model}{character string model name}
}
\value{
DE function corresponding to specific model
}
\description{
Function to select DE given model name
}
\examples{
# basic usage of hmde_model_des
hmde_model_des("constant_single_ind")

}
