% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_names.R
\name{hmde_model_name}
\alias{hmde_model_name}
\title{Returns names of available models.}
\usage{
hmde_model_name()
}
\value{
vector of character strings for model names.
}
\description{
Returns names of available models.
}
