% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_pars.R
\name{hmde_model_pars}
\alias{hmde_model_pars}
\title{Show parameter list for hmde supported model}
\usage{
hmde_model_pars(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Show parameter list for hmde supported model
}
