% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_power_de}
\alias{hmde_power_de}
\title{Differential equation for power law growth single and multi- individual models}
\usage{
hmde_power_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real}

\item{pars}{list of parameters coefficient, power}
}
\value{
value of differential equation at y
}
\description{
Differential equation for power law growth single and multi- individual models
}
