% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_run.R
\name{hmde_run}
\alias{hmde_run}
\title{Run chosen pre-built model in Stan}
\usage{
hmde_run(data_template, ...)
}
\arguments{
\item{data_template}{hmde_data_template class object}

\item{...}{additional arguments passed to rstan::sampling}
}
\value{
Stanfit S4 object
}
\description{
Run chosen pre-built model in Stan
}
\examples{
# basic usage of hmde_run
hmde_data_template("constant_single_ind",
                   obs_data = Trout_Size_Data[1:4,]) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE)

}
