% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_vb_de}
\alias{hmde_vb_de}
\title{Differential equation for von Bertalanffy growth single and multi- individual models}
\usage{
hmde_vb_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real}

\item{pars}{list of parameters Y_max, growth_rate}
}
\value{
value of differential equation at y
}
\description{
Differential equation for von Bertalanffy growth single and multi- individual models
}
