% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model_des.R
\name{hmde_zeide_de}
\alias{hmde_zeide_de}
\title{Differential equation for Zeide YD growth single and multi-individual models}
\usage{
hmde_zeide_de(y = NULL, pars = NULL)
}
\arguments{
\item{y}{input real (size)}

\item{pars}{list of parameters a (growth at 1-cm), b (growth expansion factor), c (growth decline factor)}
}
\value{
value of differential equation at y
}
\description{
Differential equation for Zeide YD growth single and multi-individual models
}
