% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{individual_ests}
\alias{individual_ests}
\alias{individual_ests<-}
\title{generic individual_ests getter}
\usage{
individual_ests(x)

individual_ests(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{tibble of individaul-level parameter estimates}
}
\description{
generic individual_ests getter

generic individual-ests setter
}
