% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{individual_ests<-,hmde_estimates-method}
\alias{individual_ests<-,hmde_estimates-method}
\title{Setter for individual_ests in hmde_estimates object}
\usage{
\S4method{individual_ests}{hmde_estimates}(x) <- value
}
\description{
Setter for individual_ests in hmde_estimates object
}
