% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{inits}
\alias{inits}
\alias{inits<-}
\title{generic inits setter}
\usage{
inits(x)

inits(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{list}
}
\description{
generic inits setter

generic inits setter
}
