% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lizard_Mass_Data}
\alias{Lizard_Mass_Data}
\title{Skink size data - Lampropholis delicata}
\format{
\subsection{\code{Lizard_Mass_Data}}{

A data frame with 336 rows and 4 columns:
\describe{
\item{ind_id}{ID number for individual}
\item{time}{Days since first observation.}
\item{y_obs}{Individual mass in kilograms.}
\item{obs_index}{Index of observations for individual}
}
}
}
\source{
\url{https://www.nature.com/articles/s41437-023-00660-3}
}
\usage{
Lizard_Mass_Data
}
\description{
A subset of data from Kar, Nakagawa, and Noble (2024), used to model growth
behaviour in a skink species.
Data was prepared by taking a simple random sample with replacement of 50
individual IDs among individuals with at least 5 observations each. Data was
then transformed to conform to the needs of a model data set in the package.
}
\keyword{datasets}
