% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lizard_Metabolic_Data}
\alias{Lizard_Metabolic_Data}
\title{Skink metabolic rate data - Lampropholis delicata}
\format{
\subsection{\code{Lizard_Metabolic_Data}}{

A data frame with 443 rows and 4 columns:
\describe{
\item{ind_id}{ID number for individual}
\item{time}{Days since first observation.}
\item{y_obs}{CO2 metabolic rate.}
\item{obs_index}{Index of observations for individual}
}
}
}
\source{
\url{https://doi.org/10.1111/oik.08122}
}
\usage{
Lizard_Metabolic_Data
}
\description{
A subset of data from Kar et al. (2021), used to model metabolic rate
variation over time across individuals.
Data was prepared by taking a simple random sample with replacement of 30
individual IDs. As two measurements were taken per observation the y_obs
used is the arithmetic mean. Every second observation was chosen in order to
moderate the data size. Data was then transformed to conform to
the needs of a model data set in the package.
}
\keyword{datasets}
