% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{measurement_ests}
\alias{measurement_ests}
\alias{measurement_ests<-}
\title{generic measurement_ests getter}
\usage{
measurement_ests(x)

measurement_ests(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{tibble of measurement-level estimates}
}
\description{
generic measurement_ests getter

generic measurement_ests setter
}
