% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{method}
\alias{method}
\alias{method<-}
\title{generic method getter}
\usage{
method(x)

method(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{character string name of object}
}
\description{
generic method getter

generic method setter
}
