% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{method,hmde_estimates-method}
\alias{method,hmde_estimates-method}
\title{Getter for method in hmde_estimates object}
\usage{
\S4method{method}{hmde_estimates}(x)
}
\description{
Getter for method in hmde_estimates object
}
