% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{.MISC}
\alias{.MISC}
\alias{.MISC<-}
\title{generic .MISC setter}
\usage{
.MISC(x)

.MISC(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{environment}
}
\description{
generic .MISC setter

generic .MISC setter
}
