% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{mode}
\alias{mode}
\alias{mode<-}
\title{generic mode setter}
\usage{
mode(x)

mode(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{integer vector}
}
\description{
generic mode setter

generic mode setter
}
