% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{model_level}
\alias{model_level}
\alias{model_level<-}
\alias{obs_data}
\alias{obs_data<-}
\title{generic model_level setter}
\usage{
model_level(x)

model_level(x) <- value

obs_data(x)

obs_data(x) <- value
}
\arguments{
\item{x}{hmde hmde_data_template}

\item{value}{character string}
}
\description{
generic model_level setter

generic model_level setter

generic obs_data setter

generic obs_data setter
}
