% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{model_name}
\alias{model_name}
\alias{model_name<-}
\title{generic model_name getter}
\usage{
model_name(x)

model_name(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{character string}
}
\description{
generic model_name getter

generic model_name setter
}
