% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{model_name,hmde_estimates-method}
\alias{model_name,hmde_estimates-method}
\title{Getter for model_name in hmde_estimates object}
\usage{
\S4method{model_name}{hmde_estimates}(x)
}
\description{
Getter for model_name in hmde_estimates object
}
