% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{model_name<-,hmde_data_template-method}
\alias{model_name<-,hmde_data_template-method}
\title{Setter for model_name in hmde_data_template object}
\usage{
\S4method{model_name}{hmde_data_template}(x) <- value
}
\description{
Setter for model_name in hmde_data_template object
}
