% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{model_pars}
\alias{model_pars}
\alias{model_pars<-}
\title{generic model_pars setter}
\usage{
model_pars(x)

model_pars(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{character vector}
}
\description{
generic model_pars setter

generic model_pars setter
}
