% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{par_dims}
\alias{par_dims}
\alias{par_dims<-}
\title{generic par_dims setter}
\usage{
par_dims(x)

par_dims(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{list}
}
\description{
generic par_dims setter

generic par_dims setter
}
