% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{par_names}
\alias{par_names}
\alias{par_names<-}
\title{generic par_names getter}
\usage{
par_names(x)

par_names(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{vector of parameter names}
}
\description{
generic par_names getter

generic par_names setter
}
