% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{par_names,hmde_data_template-method}
\alias{par_names,hmde_data_template-method}
\title{Getter for par_names in hmde_data_template object}
\usage{
\S4method{par_names}{hmde_data_template}(x)
}
\arguments{
\item{x}{hmde_data_template class object}
}
\value{
vector of character strings
}
\description{
Getter for par_names in hmde_data_template object
}
