% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{plot,hmde_data_template,ANY-method}
\alias{plot,hmde_data_template,ANY-method}
\title{Plot function for hmde_data_template object}
\usage{
\S4method{plot}{hmde_data_template,ANY}(x)
}
\arguments{
\item{x}{hmde_data_template class object}
}
\description{
Plot function for hmde_data_template object
}
\examples{
# basic usage of plot
hmde_model("constant_single_ind") |> plot()

}
