% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_S3_methods.R
\name{plot.hmde_estimates}
\alias{plot.hmde_estimates}
\title{Plot function for hmde_estimates object}
\usage{
\method{plot}{hmde_estimates}(x, ...)
}
\arguments{
\item{x}{hmde_estimates class output from hmde_extract_estimates}

\item{...}{Additional argument space to conform to S3 template.}
}
\description{
Plot function for hmde_estimates object
}
\examples{
# basic usage of print
hmde_model("constant_single_ind") |>
  hmde_assign_data(Trout_Size_Data)|>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_extract_estimates(Trout_Size_Data) |>
  plot()

}
