% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_S3_methods.R
\name{plot.hmde_model_template}
\alias{plot.hmde_model_template}
\title{Plot function for hmde_model_template object}
\usage{
plot.hmde_model_template(x, ...)
}
\arguments{
\item{x}{hmde_model_template output from hmde_model}

\item{...}{Additional argument space to conform to S3 template.}
}
\description{
Plot function for hmde_model_template object
}
\examples{
# basic usage of plot
hmde_model("constant_single_ind") |> plot()

}
