% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{population_ests}
\alias{population_ests}
\alias{population_ests<-}
\title{generic population_ests setter}
\usage{
population_ests(x)

population_ests(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{tibble of population-level estimates}
}
\description{
generic population_ests setter

generic population_ests setter
}
