% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{population_ests<-,hmde_estimates-method}
\alias{population_ests<-,hmde_estimates-method}
\title{Setter for population_ests in hmde_estimates object}
\usage{
\S4method{population_ests}{hmde_estimates}(x) <- value
}
\description{
Setter for population_ests in hmde_estimates object
}
