% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{print,hmde_data_template-method}
\alias{print,hmde_data_template-method}
\title{Print function for hmde_data_template object}
\usage{
\S4method{print}{hmde_data_template}(x)
}
\arguments{
\item{x}{hmde_data_template class object}
}
\description{
Print function for hmde_data_template object
}
\examples{
# basic usage of print
hmde_data_template("constant_single_ind") |> print()

}
