% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_S3_methods.R
\name{print.hmde_estimates}
\alias{print.hmde_estimates}
\title{Print function for hmde_estimates object}
\usage{
print.hmde_estimates(x, ...)
}
\arguments{
\item{x}{hmde_estimates class output from hmde_extract_estimates}

\item{...}{other parameters used for print}
}
\description{
Print function for hmde_estimates object
}
\examples{
# basic usage of print
hmde_model("constant_single_ind") |>
  hmde_assign_data(Trout_Size_Data)|>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_extract_estimates(Trout_Size_Data) |>
  print()

}
