% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_S3_methods.R
\name{print.hmde_model_template}
\alias{print.hmde_model_template}
\title{Print function for hmde_model_template object}
\usage{
print.hmde_model_template(x, ...)
}
\arguments{
\item{x}{hmde_model_template output from hmde_model}

\item{...}{other parameters used for print}
}
\description{
Print function for hmde_model_template object
}
\examples{
# basic usage of print
hmde_model("constant_single_ind") |> print()

}
