% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{prior_pars}
\alias{prior_pars}
\alias{prior_pars<-}
\title{generic prior_pars getter}
\usage{
prior_pars(x)

prior_pars(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{list of prior parameters}
}
\description{
generic prior_pars getter

generic prior_pars setter
}
