% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{prior_pars,hmde_data_template-method}
\alias{prior_pars,hmde_data_template-method}
\title{Getter for prior_pars in hmde_data_template object}
\usage{
\S4method{prior_pars}{hmde_data_template}(x)
}
\arguments{
\item{x}{hmde_data_template class object}
}
\value{
list of named vectors
}
\description{
Getter for prior_pars in hmde_data_template object
}
