% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{prior_pars<-,hmde_data_template-method}
\alias{prior_pars<-,hmde_data_template-method}
\title{Setter for prior_pars in hmde_data_template object}
\usage{
\S4method{prior_pars}{hmde_data_template}(x) <- value
}
\description{
Setter for prior_pars in hmde_data_template object
}
