% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{prior_pars<-,hmde_estimates-method}
\alias{prior_pars<-,hmde_estimates-method}
\title{Setter for prior_pars in hmde_estimates object}
\usage{
\S4method{prior_pars}{hmde_estimates}(x) <- value
}
\description{
Setter for prior_pars in hmde_estimates object
}
