% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{runtime}
\alias{runtime}
\alias{runtime<-}
\title{generic runtime getter}
\usage{
runtime(x)

runtime(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{matrix of chains for runtimes}
}
\description{
generic runtime getter

generic runtime setter
}
