% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{show,hmde_data_template-method}
\alias{show,hmde_data_template-method}
\title{Show function for hmde_data_template object}
\usage{
\S4method{show}{hmde_data_template}(object)
}
\arguments{
\item{object}{hmde_data_template class object}
}
\description{
Show function for hmde_data_template object
}
\examples{
# basic usage of show
hmde_data_template("constant_single_ind") |> show()

}
