% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{sim}
\alias{sim}
\alias{sim<-}
\title{generic sim setter}
\usage{
sim(x)

sim(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{list}
}
\description{
generic sim setter

generic sim setter
}
