% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Small_Tree_Size_Data}
\alias{Small_Tree_Size_Data}
\title{Protium tenufolium - Barro Colorado Island data}
\format{
\subsection{\code{Small_Tree_Size_Data}}{

A data frame with 180 rows and 4 columns:
\describe{
\item{ind_id}{ID number for individual}
\item{time}{Years since first observation.}
\item{y_obs}{Individual diameter at breast height (DBH) in centimetres.}
\item{obs_index}{Index of observations for individual}
}
}
}
\source{
\url{https://doi.org/10.15146/5xcp-0d46}
}
\usage{
Small_Tree_Size_Data
}
\description{
A subset of data from the Barro Colorado Island long term forest plot
managed by the Smithsonian Tropical Research Institute (Condit et al. 2019).
Data was prepared by taking a simple random sample without replacement of
30 individual IDs from Protium tenufolium.
The sampling frame was restricted to individuals with individuals under 10cm
for the max observed DBH, 6 observations since 1990. Data was then
transformed and renamed to match the required structure
to act as demonstration for the package.
}
\references{
\url{https://doi.org/10.1002/ecy.4140}
}
\keyword{datasets}
