% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{stan_args}
\alias{stan_args}
\alias{stan_args<-}
\title{generic stan_args setter}
\usage{
stan_args(x)

stan_args(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{list}
}
\description{
generic stan_args setter

generic stan_args setter
}
