% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_fit.R
\name{stanmodel}
\alias{stanmodel}
\alias{stanmodel<-}
\title{generic stanmodel setter}
\usage{
stanmodel(x)

stanmodel(x) <- value
}
\arguments{
\item{x}{hmde special class object}

\item{value}{stanmodel object}
}
\description{
generic stanmodel setter

generic stanmodel setter
}
