% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{summary,hmde_data_template-method}
\alias{summary,hmde_data_template-method}
\title{Summary function for hmde_data_template object}
\usage{
\S4method{summary}{hmde_data_template}(object)
}
\arguments{
\item{object}{hmde_data_template class object}
}
\description{
Summary function for hmde_data_template object
}
\examples{
# basic usage of summary
hmde_data_template("constant_single_ind") |> summary()

}
