% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_S3_methods.R
\name{summary.hmde_model_template}
\alias{summary.hmde_model_template}
\title{Summary function for hmde_model_template object}
\usage{
\method{summary}{hmde_model_template}(object, ...)
}
\arguments{
\item{object}{hmde_model_template output from hmde_model}

\item{...}{other parameters used for summary}
}
\description{
Summary function for hmde_model_template object
}
\examples{
# basic usage of summary
hmde_model("constant_single_ind") |> summary()

}
