% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tree_Size_Ests}
\alias{Tree_Size_Ests}
\title{Garcinia recondita model estimates - Barro Colorado Island data}
\format{
\subsection{\code{Tree_Size_Ests}}{

A hmde_estimates object:
\describe{
\item{model_name}{A character string giving the model name - Canham with multiple individuals.}
\item{model_level}{A character string stating that the model is fit to multiple individuals.}
\item{method}{A character string stating that the model was fit with MCMC sampling.}
\item{runtime}{A matrix giving the runtime of each chain.}
\item{fit_summary}{A character string summarising the model fitting.}
\item{measurement_ests}{A tibble with 5 columns that gives information on size observations and estimates.}
\item{individual_ests}{A tibble with 13 columns that gives posterior estimates for individual growth parameters.}
\item{population_ests}{A tibble with 5 columns that gives posterior estimates for population-level hyper-parameters.}
\item{error_ests}{A tibble with 5 columns that gives posterior estimates of the error parameter.}
\item{prior_pars}{A list contianing details of the prior parameters used for the model fit.}
\item{par_names}{A list containing the names of estimated parameters at each level of the hierarchical model.}
}
}
}
\usage{
Tree_Size_Ests
}
\description{
Estimates object of class hmde_estimates for Canham fit. Contains estimated
sizes, individual growth parameters, and population-level hyper-parameters
for Garcinia recondita fit with a Canham growth function hierarchical model.
The data used to fit the model is the Tree_Size_Data object.
}
\keyword{datasets}
