% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trout_Size_Data}
\alias{Trout_Size_Data}
\title{SUSTAIN Salmo trutta data}
\format{
\subsection{\code{Trout_Size_Data}}{

A data frame with 135 rows and 4 columns:
\describe{
\item{ind_id}{ID number for individual}
\item{time}{Years since first capture and tagging of individual.}
\item{y_obs}{Individual length in centimetres.}
\item{obs_index}{Index of observations for individual}
}
}
}
\source{
\url{https://doi.org/10.3897/BDJ.8.e52157}
}
\usage{
Trout_Size_Data
}
\description{
A subset of data from the SUSTAIN trout capture-recapture data set
from Moe et al. (2020).
Data prepared by taking a stratified sample of individual IDs based on the
number of observations per individual: 25 individuals with 2 observations,
15 with 3, 10 with 4. Within the groups a simple random sample
without replacement was used. Data was then transformed and
renamed to match the required structure to act as demonstration for
the package.
}
\keyword{datasets}
