% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{error_ests<-,hmde_estimates-method}
\alias{error_ests<-,hmde_estimates-method}
\title{Setter for error_ests in hmde_estimates object}
\usage{
\S4method{error_ests}{hmde_estimates}(x) <- value
}
\description{
Setter for error_ests in hmde_estimates object
}
