% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{fit_summary,hmde_estimates-method}
\alias{fit_summary,hmde_estimates-method}
\title{Getter for fit_summary in hmde_estimates object}
\usage{
\S4method{fit_summary}{hmde_estimates}(x)
}
\arguments{
\item{x}{hmde_estimates class object}
}
\value{
character string
}
\description{
Getter for fit_summary in hmde_estimates object
}
