% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{hmde_data_template}
\alias{hmde_data_template}
\title{Constructor function for hmde_data_template class, with data assignment ready for model fitting.}
\usage{
hmde_data_template(
  model_name,
  obs_data = NULL,
  prior_pars = NULL,
  model_level = NA_character_,
  ...
)
}
\arguments{
\item{model_name}{character string name of a hmde model}

\item{obs_data}{list or tibble containing observational data vectors}

\item{prior_pars}{list containing prior parameters}

\item{model_level}{character string specifying whether single or multiple inds}

\item{...}{data-masking name-value pairs allowing specific input of elements}
}
\value{
hmde_data_template class object
}
\description{
Constructor function for hmde_data_template class, with data assignment ready for model fitting.
}
\examples{
# basic usage of hmde_data_template
hmde_data_template("constant_single_ind")

# basic usage of hmde_data_template with data input
hmde_data_template("constant_single_ind",
  obs_data = list(
    n_obs = 3,
    y_obs = c(1,1,1),
    obs_index = 1:3,
    time = 0:2
  )
)

#basic usage of hmde_data_template with prior input
hmde_data_template("constant_single_ind",
  prior_pars = list(
    prior_pars_ind_beta = c(1,2),
    prior_pars_global_error_sigma = c(1,3)
  )
)

}
