% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{hmde_data_template}
\alias{hmde_data_template}
\title{Constructor function for hmde_data_template class, with data assignment ready for model fitting.}
\usage{
hmde_data_template(
  model_name,
  model_level = NA_character_,
  obs_data = NULL,
  prior_pars = NULL,
  ...
)
}
\arguments{
\item{model_name}{character string name of a hmde model}

\item{model_level}{character string specifying whether single or multiple inds}

\item{obs_data}{list containing observational data vectors}

\item{prior_pars}{list containing prior parameters}

\item{...}{data-masking name-value pairs allowing specific input of elements}
}
\value{
hmde_data_template class object
}
\description{
Constructor function for hmde_data_template class, with data assignment ready for model fitting.
}
\examples{
# basic usage of hmde_data_template
hmde_data_template("constant_single_ind")

}
