% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{hmde_estimates}
\alias{hmde_estimates}
\title{Constructor function for hmde_estimates class.}
\usage{
hmde_estimates(fit, obs_data)
}
\arguments{
\item{fit}{stanfit class object}

\item{obs_data}{tbl_df class object with variables for ind_id, time, y_obs}
}
\value{
hmde_estimates class object
}
\description{
Constructor function for hmde_estimates class.
}
\examples{
# basic usage of hmde_estimates
hmde_data_template("constant_single_ind",
                   obs_data = Trout_Size_Data[1:4,]) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_estimates(obs_data = Trout_Size_Data[1:4,])

}
