% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_extract_estimates.R
\name{hmde_extract_estimates}
\alias{hmde_extract_estimates}
\title{Extract samples and return measurement, individual, and population-level estimates}
\usage{
hmde_extract_estimates(fit = NULL, input_measurement_data = NULL)
}
\arguments{
\item{fit}{fitted model Stan fit}

\item{input_measurement_data}{data used to fit the model with ind_id, y_obs, time, obs_index tibble}
}
\value{
named list with data frames for measurement, individual, population-level, and error parameter estimates
}
\description{
Extract samples and return measurement, individual, and population-level estimates
}
