% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_model.R
\name{hmde_model}
\alias{hmde_model}
\title{Select data configuration template for hmde supported model}
\usage{
hmde_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Select data configuration template for hmde supported model
}
\examples{
# basic usage of hmde_model
hmde_model("constant_single_ind")

}
