% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmot_models.R
\name{hmde_model}
\alias{hmde_model}
\title{Select data configuration template for hmde supported model}
\usage{
hmde_model(model = NULL)
}
\arguments{
\item{model}{model name character string}
}
\value{
named list that matches Stan model parameters
}
\description{
Select data configuration template for hmde supported model
}
