% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_plot_de_pieces.R
\name{hmde_plot_de_pieces}
\alias{hmde_plot_de_pieces}
\title{Plot pieces of chosen differential equation model for each individual.
Function piece will go from the first fitted size to the last.
Accepted ggplot arguments will change the axis labels, title, line colour, alpha}
\usage{
hmde_plot_de_pieces(
  model = NULL,
  individual_data = NULL,
  measurement_data = NULL,
  xlab = "Y(t)",
  ylab = "f",
  title = NULL,
  colour = "#003300",
  alpha = 0.2
)
}
\arguments{
\item{model}{model name character string}

\item{individual_data}{tibble with estimated DE parameters}

\item{measurement_data}{tibble with estimated measurements}

\item{xlab}{character string for replacement x axis label}

\item{ylab}{character string for replacement y axis label}

\item{title}{character string for replacement plot title}

\item{colour}{character string for replacement line colour}

\item{alpha}{real number for replacement alpha value}
}
\value{
ggplot object
}
\description{
Plot pieces of chosen differential equation model for each individual.
Function piece will go from the first fitted size to the last.
Accepted ggplot arguments will change the axis labels, title, line colour, alpha
}
