% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_plot_de_pieces.R
\name{hmde_plot_de_pieces}
\alias{hmde_plot_de_pieces}
\title{Plot pieces of chosen differential equation model for each individual.
Structured to take the individual data tibble that is built by the
hmde_extract_estimates function using the ind_par_name_mean estimates.
Function piece will go from the first fitted size to the last.
Accepted ggplot arguments will change the axis labels, title, line colour, alpha}
\usage{
hmde_plot_de_pieces(
  estimate_list = NULL,
  xlab = "Y(t)",
  ylab = "f",
  title = NULL,
  colour = "#006600",
  alpha = 0.4
)
}
\arguments{
\item{estimate_list}{list output from hmde_extract_estimates}

\item{xlab}{character string for replacement x axis label}

\item{ylab}{character string for replacement y axis label}

\item{title}{character string for replacement plot title}

\item{colour}{character string for replacement line colour}

\item{alpha}{real number for replacement alpha value}
}
\value{
ggplot object
}
\description{
Plot pieces of chosen differential equation model for each individual.
Structured to take the individual data tibble that is built by the
hmde_extract_estimates function using the ind_par_name_mean estimates.
Function piece will go from the first fitted size to the last.
Accepted ggplot arguments will change the axis labels, title, line colour, alpha
}
\examples{
# basic usage of hmde_plot_de_pieces
hmde_plot_de_pieces(estimate_list = Tree_Size_Ests)

}
