% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_plot_Rhat_hist.R
\name{hmde_plot_Rhat_hist}
\alias{hmde_plot_Rhat_hist}
\title{Plot histogram of R_hat values for hmde_fit object.}
\usage{
hmde_plot_Rhat_hist(fit)
}
\arguments{
\item{fit}{hmde_fit object output from hmde_run}
}
\value{
ggplot object
}
\description{
Plot histogram of R_hat values for hmde_fit object.
}
\examples{
# basic usage of hmde_plot_Rhat_hist
hmde_model("constant_single_ind") |>
  hmde_assign_data(Trout_Size_Data)|>
  hmde_run(chains = 2, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_plot_Rhat_hist()

}
