% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{measurement_ests<-,hmde_estimates-method}
\alias{measurement_ests<-,hmde_estimates-method}
\title{Setter for measurement_ests in hmde_estimates object}
\usage{
\S4method{measurement_ests}{hmde_estimates}(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{tibble of estimates}
}
\value{
hmde_estimates class object
}
\description{
Setter for measurement_ests in hmde_estimates object
}
