% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{model_level,hmde_data_template-method}
\alias{model_level,hmde_data_template-method}
\title{Getter for model_level in hmde_data_template object}
\usage{
\S4method{model_level}{hmde_data_template}(x)
}
\arguments{
\item{x}{hmde_data_template class object}
}
\value{
character string
}
\description{
Getter for model_level in hmde_data_template object
}
