% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{model_level<-,hmde_data_template-method}
\alias{model_level<-,hmde_data_template-method}
\title{Setter for model_level in hmde_data_template object}
\usage{
\S4method{model_level}{hmde_data_template}(x) <- value
}
\arguments{
\item{x}{hmde_data_template class object}

\item{value}{character string}
}
\value{
hmde_data_template class object
}
\description{
Setter for model_level in hmde_data_template object
}
