% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{model_level<-,hmde_estimates-method}
\alias{model_level<-,hmde_estimates-method}
\title{Setter for model_level in hmde_estimates object}
\usage{
\S4method{model_level}{hmde_estimates}(x) <- value
}
\arguments{
\item{x}{hmde_estimates class object}

\item{value}{character string}
}
\value{
hmde_estimates class object
}
\description{
Setter for model_level in hmde_estimates object
}
