% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{obs_data,hmde_data_template-method}
\alias{obs_data,hmde_data_template-method}
\title{Getter for obs_data in hmde_data_template object}
\usage{
\S4method{obs_data}{hmde_data_template}(x)
}
\description{
Getter for obs_data in hmde_data_template object
}
