% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_data_template.R
\name{par_names<-,hmde_data_template-method}
\alias{par_names<-,hmde_data_template-method}
\title{Setter for par_names in hmde_data_template object}
\usage{
\S4method{par_names}{hmde_data_template}(x) <- value
}
\description{
Setter for par_names in hmde_data_template object
}
