% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{par_names<-,hmde_estimates-method}
\alias{par_names<-,hmde_estimates-method}
\title{Setter for par_names in hmde_estimates object}
\usage{
\S4method{par_names}{hmde_estimates}(x) <- value
}
\description{
Setter for par_names in hmde_estimates object
}
