% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{plot,hmde_estimates,ANY-method}
\alias{plot,hmde_estimates,ANY-method}
\title{Plot function for hmde_estimates object}
\usage{
\S4method{plot}{hmde_estimates,ANY}(x)
}
\arguments{
\item{x}{hmde_estimates class object}
}
\description{
Plot function for hmde_estimates object
}
\examples{
# basic usage of print
hmde_data_template("constant_single_ind",
  obs_data = Trout_Size_Data) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_estimates(Trout_Size_Data) |>
  plot()

}
