% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{runtime<-,hmde_estimates-method}
\alias{runtime<-,hmde_estimates-method}
\title{Setter for runtime in hmde_estimates object}
\usage{
\S4method{runtime}{hmde_estimates}(x) <- value
}
\description{
Setter for runtime in hmde_estimates object
}
