% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{show,hmde_estimates-method}
\alias{show,hmde_estimates-method}
\title{Show function for hmde_estimates object}
\usage{
\S4method{show}{hmde_estimates}(object)
}
\arguments{
\item{object}{hmde_estimates class output}
}
\description{
Show function for hmde_estimates object
}
\examples{
# basic usage of show
hmde_data_template("constant_single_ind",
  obs_data = Trout_Size_Data) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_estimates(Trout_Size_Data) |>
  show()

}
