% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmde_estimates.R
\name{summary,hmde_estimates-method}
\alias{summary,hmde_estimates-method}
\title{Summary function for hmde_estimates object}
\usage{
\S4method{summary}{hmde_estimates}(object)
}
\arguments{
\item{object}{hmde_estimates class object}
}
\description{
Summary function for hmde_estimates object
}
\examples{
# basic usage of summary
hmde_data_template("constant_single_ind",
  obs_data = Trout_Size_Data[1:4,]) |>
  hmde_run(chains = 1, iter = 1000,
           verbose = FALSE, show_messages = FALSE) |>
  hmde_estimates(Trout_Size_Data[1:4,]) |>
  summary()

}
