% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tree_Size_Ests}
\alias{Tree_Size_Ests}
\title{Garcinia recondita model estimates - Barro Colorado Island data}
\format{
\subsection{\code{Tree_Size_Ests}}{

A list with 4 elements:
\describe{
\item{measurement_data}{A tibble with 5 columns that gives information on size observations and estimates.}
\item{individual_data}{A tibble with 13 columns that gives posterior estimates for individual growth parameters.}
\item{error_data}{A tibble with 5 columns that gives posterior estimates of the error parameter.}
\item{population_data}{A tibble with 5 columns that gives posterior estimates for population-level hyper-parameters.}
}
}
}
\usage{
Tree_Size_Ests
}
\description{
Estimated sizes, individual growth parameters, and population-level hyper-parameters
for Garcinia recondita fit with a Canham growth function hierarchical model.
The data used to fit the model is the Tree_Size_Data object.
}
\keyword{datasets}
